Extract layout logic from FormQuickGitRefSelector into class FormQuickItemSelector
