+++ GitUI/UserControls/RevisionGrid/FormQuickItemSelector.resx
